+++ crates/holochain/src/core/workflow/sys_validation_workflow/tests.rs
@@ -21,14 +21,11 @@ async fn sys_validation_workflow_test() {
 
     let (dna_file, _, _) = SweetDnaFile::unique_from_test_wasms(vec![TestWasm::Create]).await;
 
-    let config = SweetConductorConfig::standard();
-    let mut conductors = SweetConductorBatch::from_config(2, config).await;
+    let mut conductors = SweetConductorBatch::from_standard_config_rendezvous(2).await;
     let apps = conductors.setup_app("test_app", [&dna_file]).await.unwrap();
     let ((alice,), (bob,)) = apps.into_tuples();
     let alice_cell_id = alice.cell_id().clone();
     let bob_cell_id = bob.cell_id().clone();
-
-    conductors.exchange_peer_info().await;
 
     run_test(alice_cell_id, bob_cell_id, conductors, dna_file).await;
 }
@@ -112,7 +109,7 @@ async fn sys_validation_produces_forked_chain_warrant() {
         SweetDnaFile::unique_from_inline_zomes(crate::test_utils::inline_zomes::simple_crud_zome())
             .await;
 
-    let mut conductors = SweetConductorBatch::from_standard_config(2).await;
+    let mut conductors = SweetConductorBatch::from_standard_config_rendezvous(2).await;
     let ((alice,), (bob,)) = conductors
         .setup_app("app", [&dna])
         .await