+++ crates/holochain/src/sweettest/sweet_conductor_config.rs
@@ -1,4 +1,4 @@
-use super::{DynSweetRendezvous, SweetConductor};
+use super::DynSweetRendezvous;
 use holochain_conductor_api::{
     conductor::{ConductorConfig, ConductorTuningParams, NetworkConfig},
     AdminInterfaceConfig, InterfaceDriver,
@@ -8,24 +8,19 @@ use std::sync::atomic::AtomicUsize;
 
 pub(crate) static NUM_CREATED: AtomicUsize = AtomicUsize::new(0);
 
-/// Wrapper around ConductorConfig with some helpful builder methods
+/// Wrapper around [`ConductorConfig`] with some helpful builder methods, setting
+/// default values for testing.
 #[derive(Clone, derive_more::Deref, derive_more::DerefMut, derive_more::Into)]
-pub struct SweetConductorConfig {
+pub struct SweetConductorConfig(
     #[deref]
     #[deref_mut]
     #[into]
-    config: ConductorConfig,
-
-    // Helps to keep owned references alive
-    rendezvous: Option<DynSweetRendezvous>,
-}
+    ConductorConfig,
+);
 
 impl From<ConductorConfig> for SweetConductorConfig {
     fn from(config: ConductorConfig) -> Self {
-        Self {
-            config,
-            rendezvous: None,
-        }
+        Self(config)
     }
 }
 
@@ -56,9 +51,41 @@ impl From<NetworkConfig> for SweetConductorConfig {
 }
 
 impl SweetConductorConfig {
-    /// Rewrite the config to point to the given rendezvous server
+    /// Standard config for SweetConductors.
+    ///
+    /// Bootstrapping as well as infrastructure to establish direct connections is
+    /// configured to be pointed to a locally running rendezvous server.
+    pub fn standard() -> Self {
+        let mut network_config = NetworkConfig::default()
+            .with_gossip_initiate_interval_ms(1000)
+            .with_gossip_initiate_jitter_ms(100)
+            .with_gossip_min_initiate_interval_ms(1000)
+            .with_gossip_round_timeout_ms(10_000);
+
+        network_config.bootstrap_url = url2::url2!("rendezvous:");
+        network_config.signal_url = url2::url2!("rendezvous:");
+        network_config.relay_url = url2::url2!("rendezvous:");
+
+        SweetConductorConfig::from(network_config).tune_conductor(|tune| {
+            tune.sys_validation_retry_delay = Some(std::time::Duration::from_secs(1));
+        })
+    }
+
+    /// Config for SweetConductors with a bootstrap parameter to enable or disable
+    /// bootstrapping.
+    pub fn rendezvous(bootstrap: bool) -> Self {
+        let mut config = Self::standard();
+
+        if !bootstrap {
+            config.network.disable_bootstrap = true;
+        }
+
+        config
+    }
+
+    /// Rewrite the config to point to the given rendezvous server's bootstrap,
+    /// signal and relay URLs.
     pub fn apply_rendezvous(mut self, rendezvous: &DynSweetRendezvous) -> Self {
-        self.rendezvous = Some(rendezvous.clone());
         let network = &mut self.network;
 
         if network.bootstrap_url.as_str() == "rendezvous:" {
@@ -74,43 +101,6 @@ impl SweetConductorConfig {
         }
 
         self
-    }
-
-    /// Standard config for SweetConductors
-    pub fn standard() -> Self {
-        let network_config = NetworkConfig::default()
-            .with_gossip_initiate_interval_ms(1000)
-            .with_gossip_initiate_jitter_ms(100)
-            .with_gossip_min_initiate_interval_ms(1000)
-            .with_gossip_round_timeout_ms(10_000);
-
-        SweetConductorConfig::from(network_config).tune_conductor(|tune| {
-            tune.sys_validation_retry_delay = Some(std::time::Duration::from_secs(1));
-        })
-    }
-
-    /// Rendezvous config for SweetConductors
-    pub fn rendezvous(bootstrap: bool) -> Self {
-        let mut config = Self::standard();
-
-        if bootstrap {
-            config.network.bootstrap_url = url2::url2!("rendezvous:");
-        }
-
-        config.network.signal_url = url2::url2!("rendezvous:");
-        config.network.relay_url = url2::url2!("rendezvous:");
-
-        config
-    }
-
-    /// Getter
-    pub fn get_rendezvous(&self) -> Option<DynSweetRendezvous> {
-        self.rendezvous.clone()
-    }
-
-    /// Build a SweetConductor from this config
-    pub async fn build_conductor(self) -> SweetConductor {
-        SweetConductor::from_config(self).await
     }
 
     /// Apply a function to the conductor's tuning parameters to customise them.