+++ crates/holochain/tests/tests/validate.rs
@@ -10,8 +10,7 @@ use std::sync::Arc;
 
 #[tokio::test(flavor = "multi_thread")]
 async fn call_to_validate_in_inline_zomes_passes() {
-    let config = SweetConductorConfig::standard();
-    let mut conductor = SweetConductor::from_config(config).await;
+    let mut conductor = SweetConductor::from_standard_config().await;
     let agent = SweetAgents::one(conductor.keystore()).await;
     let is_validate_called = Arc::new(AtomicBool::new(false));
     let is_validate_called_clone = is_validate_called.clone();
@@ -41,8 +40,7 @@ async fn call_to_validate_in_inline_zomes_passes() {
 
 #[tokio::test(flavor = "multi_thread")]
 async fn call_validate_across_cells_passes() {
-    let config = SweetConductorConfig::standard();
-    let mut conductor = SweetConductor::from_config(config).await;
+    let mut conductor = SweetConductor::from_standard_config().await;
     let agent = SweetAgents::one(conductor.keystore()).await;
     let is_validate_called = Arc::new(AtomicBool::new(false));
     let is_validate_called_clone = is_validate_called.clone();
@@ -91,8 +89,7 @@ async fn call_validate_across_cells_passes() {
 
 #[tokio::test(flavor = "multi_thread")]
 async fn call_validate_with_invalid_return_type_in_inline_zomes() {
-    let config = SweetConductorConfig::standard();
-    let mut conductor = SweetConductor::from_config(config).await;
+    let mut conductor = SweetConductor::from_standard_config().await;
     let agent = SweetAgents::one(conductor.keystore()).await;
     let zome = SweetInlineZomes::new(vec![], 0)
         .integrity_function("validate", |_, _: Op| Ok(42))
@@ -125,8 +122,7 @@ async fn call_validate_with_invalid_return_type_in_inline_zomes() {
 
 #[tokio::test(flavor = "multi_thread")]
 async fn call_validate_with_invalid_return_type_across_cells() {
-    let config = SweetConductorConfig::standard();
-    let mut conductor = SweetConductor::from_config(config).await;
+    let mut conductor = SweetConductor::from_standard_config().await;
     let agent = SweetAgents::one(conductor.keystore()).await;
     let zome_1 = SweetInlineZomes::new(vec![], 0)
         .integrity_function("validate", |_, _: Op| Ok(42))
@@ -175,8 +171,7 @@ async fn call_validate_with_invalid_return_type_across_cells() {
 
 #[tokio::test(flavor = "multi_thread")]
 async fn call_validate_with_invalid_parameters_in_inline_zomes() {
-    let config = SweetConductorConfig::standard();
-    let mut conductor = SweetConductor::from_config(config).await;
+    let mut conductor = SweetConductor::from_standard_config().await;
     let agent = SweetAgents::one(conductor.keystore()).await;
     let zome = SweetInlineZomes::new(vec![], 0)
         .integrity_function("validate", |_, _: usize| Ok(ValidateCallbackResult::Valid))
@@ -210,8 +205,7 @@ async fn call_validate_with_invalid_parameters_in_inline_zomes() {
 
 #[tokio::test(flavor = "multi_thread")]
 async fn call_validate_with_invalid_parameters_across_cells() {
-    let config = SweetConductorConfig::standard();
-    let mut conductor = SweetConductor::from_config(config).await;
+    let mut conductor = SweetConductor::from_standard_config().await;
     let agent = SweetAgents::one(conductor.keystore()).await;
     let zome_1 = SweetInlineZomes::new(vec![], 0)
         .integrity_function("validate", move |_, _: usize| {